requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-telegram-bot==21.7
python-dotenv==1.0.1

